Refactoring de código en clases de la historia 4
