+++ CI0126-ExamenFinal-B70866/CI0126-ExamenFinal-B70866.csproj
@@ -486,10 +486,8 @@
     </Compile>
     <Compile Include="Handlers\Handler.cs" />
     <Compile Include="Handlers\ProductHandler.cs" />
-    <Compile Include="Handlers\ShoppingCartHandler.cs" />
     <Compile Include="Helpers\Helpers.cs" />
     <Compile Include="Models\Product.cs" />
-    <Compile Include="Models\ShoppingCart.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
