+++ CI0126-ExamenFinal-B70866/Controllers/ProductController.cs
@@ -2,6 +2,7 @@
 using System.Web.Mvc;
 using CI0126_ExamenFinal_B70866.Models;
 using CI0126_ExamenFinal_B70866.Handlers;
+using System.Collections.Generic;
 
 namespace CI0126_ExamenFinal_B70866.Controllers
 {
@@ -39,20 +40,33 @@ namespace CI0126_ExamenFinal_B70866.Controllers
         [HttpGet]
         public ActionResult addProductToCart(int productID, int amount)
         {
-            ShoppingCartHandler shoppingCartHandler = new ShoppingCartHandler();
+            ProductHandler productHandler = new ProductHandler();
             Product newProduct = new Product();
             newProduct.id = productID;
             newProduct.amount = amount;
-            shoppingCartHandler.addProductToCart(newProduct);
+            productHandler.addProductToCart(newProduct);
             return RedirectToAction("shoppingCart");
+        }
+
+        public double getTotalPrice()
+        {
+            double totalPrice = 0;
+            ProductHandler productHandler = new ProductHandler();
+            List<Product> productList = productHandler.getAllProductsInCart().ToList();
+            foreach (var product in productList)
+            {
+                totalPrice += product.price;
+            }
+            return totalPrice;
         }
 
         [HandleError]
         public ActionResult shoppingCart()
         {
-            ShoppingCartHandler shoppingCartHandler = new ShoppingCartHandler();
-            ViewBag.productsInCart = shoppingCartHandler.getAllProductsInCart().ToList<Product>();
-            return View(ViewBag.productsInCart);
+            ProductHandler productHandler = new ProductHandler();
+            ViewBag.productsInCart = productHandler.getAllProductsInCart().ToList();
+            ViewBag.totalPrice = getTotalPrice();
+            return View();
         }
     }
 }
