--- CI0126-ExamenFinal-B70866/Models/ShoppingCart.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-namespace CI0126_ExamenFinal_B70866.Models
-{
-    public class ShoppingCart
-    {
-        public List<Product> products { get; set; }
-    }
-}
